requirements.txt
aiohttp
aiosqlite
discord
discord.py-self @ git+https://github.com/dolfies/discord.py-self.git
python-dotenv